+++ src/config/config_definition.cc
@@ -30,21 +30,24 @@
 #include <curl/curl.h>
 #endif
 
-#include "client_config.h"
+#include "config/result/box_layout.h"
+#include "config/result/client_config.h"
 #include "config/setup/config_setup_autoscan.h"
 #include "config/setup/config_setup_boxlayout.h"
 #include "config/setup/config_setup_client.h"
+#include "config/setup/config_setup_dictionary.h"
 #include "config/setup/config_setup_dynamic.h"
 #include "config/setup/config_setup_enum.h"
+#include "config/setup/config_setup_path.h"
 #include "config/setup/config_setup_time.h"
 #include "config/setup/config_setup_transcoding.h"
 #include "config/setup/config_setup_tweak.h"
+#include "config/setup/config_setup_vector.h"
 #include "config_option_enum.h"
 #include "config_options.h"
 #include "config_setup.h"
 #include "content/autoscan.h"
 #include "content/content_manager.h"
-#include "content/layout/box_layout.h"
 #include "database/sqlite3/sqlite_config.h"
 #include "metadata/metadata_handler.h"
 #include "upnp_common.h"
