+++ src/config/config_generator.cc
@@ -30,9 +30,11 @@
 #include "config/config_definition.h"
 #include "config/config_options.h"
 #include "config/config_setup.h"
+#include "config/result/box_layout.h"
 #include "config/setup/config_setup_boxlayout.h"
+#include "config/setup/config_setup_dictionary.h"
+#include "config/setup/config_setup_vector.h"
 #include "content/autoscan.h"
-#include "content/layout/box_layout.h"
 #include "util/grb_time.h"
 #include "util/tools.h"
