+++ src/config/result/client_config.cc
