+++ src/config/result/dynamic_content.cc
@@ -25,6 +25,7 @@
 
 #include "dynamic_content.h" // API
 
+#include "content/autoscan.h"
 #include "content/content_manager.h"
 
 void DynamicContentList::add(const std::shared_ptr<DynamicContent>& cont, std::size_t index)
