+++ src/config/setup/config_setup_array.cc
@@ -26,7 +26,6 @@
 
 #include <numeric>
 
-#include "config/client_config.h"
 #include "config/config_definition.h"
 #include "config/config_options.h"
