+++ src/config/setup/config_setup_path.cc
@@ -22,9 +22,8 @@
 
 /// \file config_setup_path.cc
 
-#include "config/config_setup.h" // API
+#include "config_setup_path.h" // API
 
-#include "config/client_config.h"
 #include "config/config_definition.h"
 #include "config/config_options.h"
