+++ src/content/scripting/script.cc
@@ -41,12 +41,12 @@
 #include "cds/cds_container.h"
 #include "cds/cds_item.h"
 #include "config/config_definition.h"
-#include "config/config_setup.h"
+#include "config/result/box_layout.h"
 #include "config/setup/config_setup_autoscan.h"
 #include "config/setup/config_setup_boxlayout.h"
+#include "config/setup/config_setup_dictionary.h"
 #include "content/autoscan.h"
 #include "content/content_manager.h"
-#include "content/layout/box_layout.h"
 #include "database/database.h"
 #include "js_functions.h"
 #include "metadata/metadata_handler.h"
