+++ src/web/config_load.cc
@@ -29,18 +29,20 @@
 #include <fmt/chrono.h>
 #include <numeric>
 
-#include "config/client_config.h"
 #include "config/config_definition.h"
 #include "config/config_option_enum.h"
 #include "config/config_setup.h"
-#include "config/directory_tweak.h"
-#include "config/dynamic_content.h"
+#include "config/result/box_layout.h"
+#include "config/result/client_config.h"
+#include "config/result/directory_tweak.h"
+#include "config/result/dynamic_content.h"
 #include "config/setup/config_setup_autoscan.h"
 #include "config/setup/config_setup_boxlayout.h"
+#include "config/setup/config_setup_dictionary.h"
 #include "config/setup/config_setup_enum.h"
+#include "config/setup/config_setup_vector.h"
 #include "content/autoscan.h"
 #include "content/content_manager.h"
-#include "content/layout/box_layout.h"
 #include "database/database.h"
 #include "metadata/metadata_handler.h"
 #include "transcoding/transcoding.h"
