+++ src/web/config_save.cc
@@ -26,7 +26,6 @@
 
 #include "pages.h" // API
 
-#include "config/client_config.h"
 #include "config/config.h"
 #include "config/config_definition.h"
 #include "config/config_setup.h"
