+++ test/util/test_upnp_clients.cc
@@ -22,7 +22,7 @@
 #include <gtest/gtest.h>
 #include <upnp.h>
 
-#include "config/client_config.h"
+#include "config/result/client_config.h"
 #include "util/grb_net.h"
 #include "util/upnp_clients.h"
